FigureSummarizer/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.25.0
Pillow>=10.0.0
PyMuPDF>=1.23.0
numpy>=1.24.0
tqdm>=4.65.0
